winter_fri::folding

Function apply_drp

pub fn apply_drp<B, E, const N: usize>(
    values: &[[E; N]],
    domain_offset: B,
    alpha: E,
) -> Vec<E>
where
    B: StarkField,
    E: FieldElement<BaseField = B>,

Applies degree-respecting projection to evaluations of a polynomial.

This process reduces the degree of the polynomial implied by the evaluations by N. For example, for N = 4, this is equivalent to the following:

Let evaluations contain the evaluations of polynomial f(x) of degree k.

Group coefficients of f so that f(x) = a(x) + x * b(x) + x^2 * c(x) + x^3 * d(x)

Compute random linear combination of polynomials a, b, c, d as f’(x) = a + α * b + α^2 * c + α^3 * d, where α is an element drawn uniformly at random from the entire field.

evaluate f’(x) on a domain which consists of x^4 from the original domain (and thus is 1/4 the size).

However, the reduction is performed without converting the polynomials into coefficient form. That is, we can go directly form evaluations to folded evaluations. For this, for each evaluation in the folded domain, we need N evaluations in the source domain. For example, for the case of N = 4, to compute the evaluation of f’(x) we need to have evaluations of f(x) from the source domain at x^{1/4}, x^{2/4}, x^{3/4}, x.

This function expect the evaluations to be already in a transposed form such that all evaluations needed to compute a single evaluation in the folded domain are next to each other.

The example below shows the equivalence of performing the projection via coefficient form and via evaluation form for N = 2.

// generate random alpha
let alpha: BaseElement = rand_value();

// degree 7 polynomial f(x)
let poly: Vec<BaseElement> = rand_vector(8);

// f'(x) = g(x) + alpha * h(x) such that g(x) consists of even coefficients of f(x) and h(x)
// consists of odd coefficients of f(x). This is equivalent to using `folding_factor = 2`.
let mut folded_poly = Vec::new();
for i in 0..poly.len() / 2 {
    folded_poly.push(poly[2 * i] + alpha * poly[2 * i + 1]);
}

// build a domain of 32 elements
let n = 32_usize;
let offset = BaseElement::GENERATOR;
let g = BaseElement::get_root_of_unity(n.trailing_zeros());
let domain = get_power_series_with_offset(g, offset, n);

// build a folded domain of half the size, such that x in the source domain maps to
// x^2 in the folded domain.
let g = BaseElement::get_root_of_unity((n / 2).trailing_zeros());
let folded_domain = get_power_series_with_offset(g, offset.exp(2), n / 2);

// evaluate the polynomials over their respective domains
let evaluations = polynom::eval_many(&poly, &domain);
let folded_evaluations = polynom::eval_many(&folded_poly, &folded_domain);

// use DRP to perform folding without converting polynomials into coefficient form
let transposed_evaluations = transpose_slice::<BaseElement, 2>(&evaluations);
let drp_evaluations = apply_drp(&transposed_evaluations, offset, alpha);

// applying DRP should be equivalent to folding polynomials in coefficient form
assert_eq!(folded_evaluations, drp_evaluations);