winter_fri::folding

Function fold_positions

pub fn fold_positions(
    positions: &[usize],
    source_domain_size: usize,
    folding_factor: usize,
) -> Vec<usize>

Maps positions in the source domain, to positions in the folded domain.

The size of the folded domain is reduced by the folding_factor as compared to the size of the source domain. Thus, the original positions may fall outside of the folded domain. To map positions in the source domain to positions in the folded domain, we simply compute new_position = old_position % folded_domain_size and discard duplicates.

let positions = [1, 9, 12, 20];
let folded_positions = fold_positions(&positions, 32, 4);

// 1 and 9 should map to 1; 12 and 20 should map to 4
assert_eq!(vec![1, 4], folded_positions);

The domains we usually work with have the following form: c, c * ω, c * ω^2, …, c * ω^(n-1), where n is the size of the domain, ω is nth root of unity, and c is the domain offset. Thus, the procedure described above maps c * ω^i to (c * ω^i)^k, where k is the folding factor.

// build a domain of size 32
let n = 32usize;
let c = BaseElement::GENERATOR;
let g = BaseElement::get_root_of_unity(n.trailing_zeros());
let source_domain = get_power_series_with_offset(g, c, n);

// build a domain of size 8 (folding_factor = 4)
let g = BaseElement::get_root_of_unity((n / 4).trailing_zeros());
let folded_domain = get_power_series_with_offset(g, c.exp(4_u32.into()), (n / 4));

// position 1 in the source domain maps to position 1 in the folded domain
assert_eq!(folded_domain[1], source_domain[1].exp(4));

// position 9 in the source domain also maps to position 1 in the folded domain
assert_eq!(folded_domain[1], source_domain[9].exp(4));