winter_fri

Struct DefaultVerifierChannel

pub struct DefaultVerifierChannel<E: FieldElement, H: ElementHasher<BaseField = E::BaseField>, V: VectorCommitment<H>> { /* private fields */ }

Provides a default implementation of the VerifierChannel trait.

Default verifier channel can be instantiated directly from a FriProof struct.

Though this implementation is primarily intended for testing purposes, it can be used in production use cases as well.

Implementations

impl<E, H, V> DefaultVerifierChannel<E, H, V>
where E: FieldElement, H: ElementHasher<BaseField = E::BaseField>, V: VectorCommitment<H>,

pub fn new( proof: FriProof, layer_commitments: Vec<H::Digest>, domain_size: usize, folding_factor: usize, ) -> Result<Self, DeserializationError>

Builds a new verifier channel from the specified FriProof.

Errors

Returns an error if the specified proof could not be parsed correctly.