winter_fri

Trait ProverChannel

pub trait ProverChannel<E: FieldElement> {
    type Hasher: ElementHasher<BaseField = E::BaseField>;

    // Required methods
    fn commit_fri_layer(&mut self, layer_root: <Self::Hasher as Hasher>::Digest);
    fn draw_fri_alpha(&mut self) -> E;
}

Defines an interface for a channel over which a prover communicates with a verifier.

The prover uses this channel to send commitments to FRI layer polynomials to the verifier, and then to draw a random value α from the channel after each commitment is sent. The prover then uses this α to construct the next FRI layer.

In the interactive version of the protocol, the verifier chooses α uniformly at random from the entire field. In the non-interactive version, the α is drawn pseudo-randomly based on the commitments the prover has written into the channel up to this point.

Required Associated Types

type Hasher: ElementHasher<BaseField = E::BaseField>

Hash function used by the prover to commit to polynomial evaluations.

Required Methods

fn commit_fri_layer(&mut self, layer_root: <Self::Hasher as Hasher>::Digest)

Sends a layer commitment to the verifier.

A layer commitment is the commitment string of a vector commitment to the vector of evaluations of a polynomial at a given layer. The vector commitment is built by first transposing evaluations into a two-dimensional matrix where each row contains values needed to compute a single value of the next FRI layer, and then computing the hash of each row to get one entry of the vector being committed to. Thus, the number of elements grouped into a single leaf is equal to the folding_factor used for FRI layer construction.

fn draw_fri_alpha(&mut self) -> E

Returns a random α drawn uniformly at random from the entire field.

The prover uses this α to build the next FRI layer.

While in the interactive version of the protocol the verifier send a random α to the prover, in the non-interactive version, the α is pseudo-randomly generated based on the values the prover previously wrote into the channel.