winter_fri

Trait VerifierChannel

pub trait VerifierChannel<E: FieldElement> {
    type Hasher: ElementHasher<BaseField = E::BaseField>;
    type VectorCommitment: VectorCommitment<Self::Hasher>;

    // Required methods
    fn read_fri_num_partitions(&self) -> usize;
    fn read_fri_layer_commitments(
        &mut self,
    ) -> Vec<<Self::Hasher as Hasher>::Digest>;
    fn take_next_fri_layer_queries(&mut self) -> Vec<E>;
    fn take_next_fri_layer_proof(
        &mut self,
    ) -> <Self::VectorCommitment as VectorCommitment<Self::Hasher>>::MultiProof;
    fn take_fri_remainder(&mut self) -> Vec<E>;

    // Provided methods
    fn read_layer_queries<const N: usize>(
        &mut self,
        positions: &[usize],
        commitment: &<Self::Hasher as Hasher>::Digest,
    ) -> Result<Vec<[E; N]>, VerifierError> { ... }
    fn read_remainder(&mut self) -> Result<Vec<E>, VerifierError> { ... }
}

Defines an interface for a channel over which a verifier communicates with a prover.

This trait abstracts away implementation specifics of the FriProof struct. Thus, instead of dealing with FRI proofs directly, the verifier can read the data as if it was sent by the prover via an interactive channel.

Note: that reading removes the data from the channel. Thus, reading duplicated values from the channel should not be possible.

Required Associated Types

type Hasher: ElementHasher<BaseField = E::BaseField>

Hash function used by the prover to commit to polynomial evaluations.

type VectorCommitment: VectorCommitment<Self::Hasher>

Vector commitment used to commit to polynomial evaluations.

Required Methods

fn read_fri_num_partitions(&self) -> usize

Returns the number of partitions used during proof generation.

fn read_fri_layer_commitments( &mut self, ) -> Vec<<Self::Hasher as Hasher>::Digest>

Reads and removes from the channel all FRI layer commitments sent by the prover.

In the interactive version of the protocol, the prover sends layer commitments to the verifier one-by-one, and the verifier responds with a value α drawn uniformly at random from the entire field after each layer commitment is received. In the non-interactive version, the verifier can read all layer commitments at once, and then generate α values locally.

fn take_next_fri_layer_queries(&mut self) -> Vec<E>

Reads and removes from the channel evaluations of the polynomial at the queried positions for the next FRI layer.

In the interactive version of the protocol, these evaluations are sent from the prover to the verifier during the query phase of the FRI protocol.

It is expected that layer queries and layer proofs at the same FRI layer are consistent. That is, query values hash into the leaf nodes of corresponding vector commitment.

fn take_next_fri_layer_proof( &mut self, ) -> <Self::VectorCommitment as VectorCommitment<Self::Hasher>>::MultiProof

Reads and removes from the channel vector commitment opening proofs of queried evaluations for the next FRI layer.

In the interactive version of the protocol, these authentication paths are sent from the prover to the verifier during the query phase of the FRI protocol.

It is expected that layer proofs and layer queries at the same FRI layer are consistent. That is, query values hash into the elements of the vector committed to using the specified vector commitment scheme.

fn take_fri_remainder(&mut self) -> Vec<E>

Reads and removes the remainder polynomial from the channel.

Provided Methods

fn read_layer_queries<const N: usize>( &mut self, positions: &[usize], commitment: &<Self::Hasher as Hasher>::Digest, ) -> Result<Vec<[E; N]>, VerifierError>

Returns FRI query values at the specified positions from the current FRI layer and advances layer pointer by one.

This also checks if the values are valid against the provided FRI layer commitment.

Errors

Returns an error if query values did not match layer commitment.

fn read_remainder(&mut self) -> Result<Vec<E>, VerifierError>

Returns FRI remainder polynomial read from this channel.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.